requirements.txt
# Trunk Player v2 - Requirements
# Python 3.12+

# Core Django
Django==5.1.7
asgiref==3.8.1
daphne==4.1.2

# Database
psycopg2-binary==2.9.10

# Redis & Channels
redis==5.0.8
hiredis==3.0.0
django-redis==5.4.0
channels==4.1.0
channels-redis==4.2.0

# REST API
djangorestframework==3.15.2
drf-spectacular==0.28.0

# Authentication - Fief
fief-client==0.19.0
httpx==0.27.0

# Background Tasks (for future transcription processing)
celery==5.4.0
django-celery-beat==2.7.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1
requests==2.32.3

# Async support
Twisted==24.7.0
txaio==23.1.1

# Security
cryptography==43.0.0
PyJWT==2.9.0

# Type annotations
typing-extensions==4.12.2
